requirements.txt
APScheduler==3.11.0
click==8.3.0
joblib==1.5.2
nltk==3.9.2
python-dotenv==1.1.1
regex==2025.9.18
secure-smtplib==0.1.1
tqdm==4.67.1
tzlocal==5.3.1
requests==2.32.3
beautifulsoup4==4.12.3
